findproperty does digital land fetch
